data_colector/requirements.txt
requests==2.10.0
websocket-client==1.2.1
psycopg2==2.9.1